+++ ChangeLog.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+
+### Added
+
+- Support for OpenAPI plugins.
+
+### Changed
+
+- JSON Patch support is now a plugin.
+
+
 ## [0.11.1] - 2024-08-15
 
 ### Removed